set up files

+++ core/submodules/Frontier/submodules/User/submodules/Role/API/Controllers/GrantController.php
@@ -56,6 +56,24 @@ class GrantController extends APIController
     }
 
     /**
+     * Get all resources.
+     *
+     * @param  Illuminate\Http\Request $request [description]
+     * @return Illuminate\Http\Response
+     */
+    public function getTrash(Request $request)
+    {
+        $search = $request->get('q') !== 'null' && $request->get('q') ? $request->get('q'): '';
+        $take = $request->get('take') && $request->get('take') > 0 ? $request->get('take') : 0;
+        $sort = $request->get('sort') && $request->get('sort') !== 'null' ? $request->get('sort') : 'id';
+        $order = $request->get('descending') === 'true' && $request->get('descending') !== 'null' ? 'DESC' : 'ASC';
+
+        $permissions = Role::search($search)->orderBy($sort, $order)->onlyTrashed()->paginate($take);
+
+        return response()->json($permissions);
+    }
+
+    /**
      * Gets the permissions.
      *
      * @param  array $modules
@@ -83,6 +101,14 @@ class GrantController extends APIController
     public function destroy(Request $request, $id)
     {
         $grant = Grant::findOrFail($id);
+
+        if (in_array($grant->code, config('auth.rootgrants', []))) {
+            $this->errorResponse['text'] = "Deleting Root Grants is not permitted";
+
+            return response()->json($this->errorResponse);
+        }
+
+        $this->successResponse['text'] = "{$grant->name} moved to trash.";
         $grant->delete();
 
         return response()->json($this->successResponse);
@@ -96,14 +122,43 @@ class GrantController extends APIController
      */
     public function clone(Request $request, $id)
     {
-        $grant = Grant::findOrFail($id);
+        $role = Grant::findOrFail($id);
 
         $clone = new Grant();
-        $clone->name = $grant->name;
-        $clone->code = "{$grant->code}-clone-{$id}";
-        $clone->description = $grant->description;
+        $clone->name = $role->name;
+        $clone->code = "{$role->code}-clone-".rand((int) $id, (int) date('Y'));
+        $clone->description = $role->description;
         $clone->save();
-        $clone->permissions()->attach($grant->permissions->pluck('id')->toArray());
+        $clone->grants()->attach($role->grants->pluck('id')->toArray());
+
+        return response()->json($this->successResponse);
+    }
+
+    /**
+     * Restore the specified resource from storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function restore(Request $request, $id)
+    {
+        $grant = Grant::onlyTrashed()->findOrFail($id);
+        $grant->restore();
+
+        return response()->json($this->successResponse);
+    }
+
+    /**
+     * Delete the specified resource from storage permanently.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function delete(Request $request, $id)
+    {
+        $grant = Role::withTrashed()->findOrFail($id);
+        $grant->forceDelete();
 
         return response()->json($this->successResponse);
     }
